회원가입 페이지 진입시 휴대폰인증 세션정보 초기화 되도록 수정. 소유권 발급 요청 및 내역을 오래된 순으로 보여주도록 정렬

+++ backend/src/utils/jwt.js
@@ -24,8 +24,10 @@ async function createLoginToken(user, userAgent) {
     const refreshToken = createToken(payload, process.env.REFRESH_TOKEN_EXPIRES);
 
     // 기존에 보관된 로그인 토큰 정보 삭제
-    sql = `DELETE FROM LoginToken WHERE username='${user.username}'`;
-    await con.execute(sql);
+    if (user.username) {
+      sql = `DELETE FROM LoginToken WHERE username='${user.username}'`;
+      await con.execute(sql);
+    }
 
     // DB에 로그인 정보와 새로운 토큰 정보 저장
     sql = `INSERT INTO
